Move omprog_give beside the other item progs
feniaroot's obj.trigger("Give") could not link against a function defined
in the comm plugin (drone build 1116). oprog_get, oprog_drop and
oprog_give_blocked already live in fight_core/item_progs.cpp, which every
caller can reach; omprog_give belongs with them.

No behaviour change: same body, now declared in item_progs.h.
