+++ plug-ins/fight_core/item_progs.h
@@ -14,4 +14,8 @@ bool oprog_drop( Object *obj, Character *ch );
 // item moves, so a veto leaves both characters exactly as they were.
 bool oprog_give_blocked( Object *obj, Character *ch, Character *victim );
 
+// Runs the full hand-over chain once the item is already in victim's hands, and
+// returns true if anything along it claimed or took back the item.
+bool omprog_give( Object *obj, Character *ch, Character *victim );
+
 #endif
